% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{ad_SI}
\alias{ad_SI}
\title{Standard air density}
\format{An object of class \code{numeric} of length 1.}
\usage{
ad_SI
}
\description{
Standard air density at METRIC standard conditions, [kg/m3]
}
\keyword{datasets}

