% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{ad_US}
\alias{ad_US}
\title{Standard air density}
\format{An object of class \code{numeric} of length 1.}
\usage{
ad_US
}
\description{
Standard air density at FIELD standard conditions, [lb/ft3]
}
\keyword{datasets}

