% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{API_gcc}
\alias{API_gcc}
\title{Density - Gravity conversion}
\usage{
API_gcc(x)
}
\arguments{
\item{x}{A quantity to convert from, [API]}
}
\value{
Converted quantity, [g/cm3]
}
\description{
Density conversion: FIELD [API] to METRIC [g/cm3] (specific gravity)
}
\examples{
API_gcc(32)
}
