% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{API_to_kgm3}
\alias{API_to_kgm3}
\title{Density conversion}
\usage{
API_to_kgm3(x)
}
\arguments{
\item{x}{A quantity to convert from, [API]}
}
\value{
Converted quantity, [kg/m3]
}
\description{
Density conversion: FIELD [API] to METRIC [kg/m3]
}
\examples{
API_to_kgm3(32)
}

