% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{atm_ata}
\alias{atm_ata}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
atm_ata
}
\description{
Pressure: atmposphere [atm] to absolute atmposphere [ata] conversion factor
}
\keyword{datasets}
