% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{atm_mmHg}
\alias{atm_mmHg}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
atm_mmHg
}
\description{
Pressure: atmposphere [atm] to Hg millimeters [mmHg] conversion factor
}
\keyword{datasets}
