% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{atm_Pa}
\alias{atm_Pa}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
atm_Pa
}
\description{
Pressure: atmposphere [atm] to Pascal [Pa] conversion factor
}
\keyword{datasets}
