% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{bar_psi}
\alias{bar_psi}
\title{Pressure conversion}
\usage{
bar_psi(x)
}
\arguments{
\item{x}{A quantity to convert from, [bar]}
}
\value{
Converted quantity, [psi]
}
\description{
Pressure conversion: METRIC [bar] to FIELD [psi]
}
\examples{
bar_psi(32)
}

