% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{bbl_ft3}
\alias{bbl_ft3}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
bbl_ft3
}
\description{
Volume: barrel [bbl] to cubic feet [ft3] conversion factor
}
\keyword{datasets}
