% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{BTU_Jmol}
\alias{BTU_Jmol}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
BTU_Jmol
}
\description{
Energy: British thermal unit [BTU] to Joule/mole [J/mole] conversion factor
}
\keyword{datasets}
