% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{C_F}
\alias{C_F}
\title{Temperature conversion}
\usage{
C_F(x)
}
\arguments{
\item{x}{A quantity to convert from, [C]}
}
\value{
Converted quantity, [F]
}
\description{
Temperature conversion: Celsius to Farenheit
}
\examples{
C_F(15.6)
}

