% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{C_R}
\alias{C_R}
\title{Temperature conversion}
\usage{
C_R(x)
}
\arguments{
\item{x}{A quantity to convert from, [C]}
}
\value{
Converted quantity, [R]
}
\description{
Temperature conversion: Celsius to Rankine
}
\examples{
C_R(15.6)
}
