% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{day_s}
\alias{day_s}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
day_s
}
\description{
Time: day to second conversion
}
\keyword{datasets}
