% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{F_C}
\alias{F_C}
\title{Temperature conversion}
\usage{
F_C(x)
}
\arguments{
\item{x}{A quantity to convert from, [F]}
}
\value{
Converted quantity, [C]
}
\description{
Temperature conversion: Farenheit to Celsius
}
\examples{
F_C(60)
}

