% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{ft_m}
\alias{ft_m}
\title{Important conversion factors apart from unit lists}
\format{An object of class \code{numeric} of length 1.}
\usage{
ft_m
}
\description{
Length: foot [ft] to [m] conversion factor
}
\keyword{datasets}

