% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{gcc_API}
\alias{gcc_API}
\title{Density - Gravity conversion}
\usage{
gcc_API(x)
}
\arguments{
\item{x}{A quantity to convert from, [g/cm3]}
}
\value{
Converted quantity, [API]
}
\description{
Density conversion: METRIC [g/cm3] (specific gravity) to FIELD [API]
}
\examples{
gcc_API(0.865)
}
