% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{GLR_SI}
\alias{GLR_SI}
\title{Gas-liquid ratio conversion}
\usage{
GLR_SI(x)
}
\arguments{
\item{x}{A quantity to convert from (numeric vector), [scf/STB]}
}
\value{
Converted quantity (numeric vector), [m3/m3]
}
\description{
Gas-liquid ratio conversion: FIELD [scf/STB] to METRIC [m3/m3]
}
\examples{
 GLR_SI(c(500, 1100))
}
