% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{GLR_US}
\alias{GLR_US}
\title{Gas-liquid ratio conversion}
\usage{
GLR_US(x)
}
\arguments{
\item{x}{A quantity to convert from (numeric vector), [m3/m3]}
}
\value{
Converted quantity (numeric vector), [scf/STB]
}
\description{
Gas-liquid ratio conversion: METRIC [m3/m3] to FIELD [scf/STB]
}
\examples{
 GLR_US(c(85, 110))
}
