% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{is.physical}
\alias{is.physical}
\title{Generic function: check physical measure}
\usage{
is.physical(x)
}
\arguments{
\item{x}{A quantity to check}
}
\value{
TRUE or FALSE whether x is an object of class "physical"
}
\description{
Generic function: check physical measure
}
\examples{
 is.physical(32)
 is.physical(u(32, "deg.C"))
}

