% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{kgm3_API}
\alias{kgm3_API}
\title{Density - Gravity conversion}
\usage{
kgm3_API(x)
}
\arguments{
\item{x}{A quantity to convert from, [kg/m3]}
}
\value{
Converted quantity, [API]
}
\description{
Density conversion: METRIC [kg/m3] to FIELD [API]
}
\examples{
kgm3_API(865)
}
