% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{lb_kg}
\alias{lb_kg}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
lb_kg
}
\description{
Mass: pound [lb] to [kg] conversion factor
}
\keyword{datasets}
