% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{mD_m2}
\alias{mD_m2}
\title{Important conversion factors apart from unit lists}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
mD_m2
}
\description{
Area: millidarcy [mD] to square meter [m2] conversion factor
}
\keyword{datasets}
