% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{MPa_psi}
\alias{MPa_psi}
\title{Pressure conversion}
\usage{
MPa_psi(x)
}
\arguments{
\item{x}{A quantity to convert from, [MPa]}
}
\value{
Converted quantity, [psi]
}
\description{
Pressure conversion: METRIC [MPa] to FIELD [psi]
}
\examples{
MPa_psi(32)
}
