% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{MWair}
\alias{MWair}
\title{Built-in constant: Air molecular mass}
\format{An object of class \code{numeric} of length 1.}
\usage{
MWair
}
\description{
Air molecular mass, 28.967 [g/mole]
}
\keyword{datasets}

