% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{Pressure}
\alias{Pressure}
\title{Pressure unit list}
\usage{
Pressure()
}
\description{
Pressure unit list
}
\details{
Valid units: "mPa", "Pa", "kPa", "MPa", "GPa",
 "psi", "atm", "ata" (absolute atmosphere), "bar" and "mmHg"
}

