% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{print.physical}
\alias{print.physical}
\title{Generic function: print physical measure}
\usage{
\method{print}{physical}(x, ...)
}
\arguments{
\item{x}{A quantity to print}

\item{...}{Optional arguments to be passed further}
}
\value{
Prints to console x with "Units: " at the beginning of the object output
}
\description{
Generic function: print physical measure
}
\examples{
 z <- round(runif(10)*100, 1)
 print(u(z, "mD"))
}
