% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{psi_bar}
\alias{psi_bar}
\title{Pressure conversion}
\usage{
psi_bar(x)
}
\arguments{
\item{x}{A quantity to convert from, [psi]}
}
\value{
Converted quantity, [bar]
}
\description{
Pressure conversion: FIELD [psi] to METRIC [bar]
}
\examples{
psi_bar(32)
}
