% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{psi_MPa}
\alias{psi_MPa}
\title{Pressure conversion}
\usage{
psi_MPa(x)
}
\arguments{
\item{x}{A quantity to convert from, [psi]}
}
\value{
Converted quantity, [MPa]
}
\description{
Pressure conversion: FIELD [psi] to METRIC [MPa]
}
\examples{
psi_MPa(32)
}
