% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{psi_Pa}
\alias{psi_Pa}
\title{Important conversion factors apart from unit lists}
\format{An object of class \code{numeric} of length 1.}
\usage{
psi_Pa
}
\description{
Pressure: psi [psi]=[lb/in2] to Pascal [Pa] conversion factor
}
\keyword{datasets}

