% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{R_SI}
\alias{R_SI}
\title{Universal gas constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
R_SI
}
\description{
Universal gas constant in SI units, [J/(mole*K)]
}
\keyword{datasets}

