% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{R_to_C}
\alias{R_to_C}
\title{Temperature conversion}
\usage{
R_to_C(x)
}
\arguments{
\item{x}{A quantity to convert from, [F]}
}
\value{
Converted quantity, [C]
}
\description{
Temperature conversion: Rankine to Celsius
}
\examples{
R_to_C(60 + TF_TR)
}

