% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{TC_TK}
\alias{TC_TK}
\title{Celsius to Kelvin temperature constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
TC_TK
}
\description{
Celsius to Kelvin temperature constant
}
\keyword{datasets}

