% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{TF_TR}
\alias{TF_TR}
\title{Rankine to Farenheit temperature constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
TF_TR
}
\description{
Rankine to Farenheit temperature constant
}
\keyword{datasets}

