% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{u}
\alias{u}
\title{Unit create function}
\usage{
u(x, unit)
}
\arguments{
\item{x}{A quantity to use as physical meaning value}

\item{unit}{Valid unit name}
}
\value{
Object of class "physical"
}
\description{
Creates a class with specified unit
}
\examples{
 z1 <- u(1:5, "km")
 print(z1)
 z2 <- u(list(1, 2), "kg")
 print(z2)
}
