% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{un}
\alias{un}
\title{Unit destroy function}
\usage{
un(x)
}
\arguments{
\item{x}{A quantity of class "physical"}
}
\value{
Dimensionless vector or list
}
\description{
Removes units from vector of list
}
\examples{
 z1 <- u(list(1, 2, 3), "km")
 print(z1)
 z2 <- un(z1)
 print(z2)
}

