% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\name{Vm_SI}
\alias{Vm_SI}
\title{Molar volume in METRIC units}
\usage{
Vm_SI()
}
\description{
Molar volume in METRIC units [l] per 1 g-mole
 at FIELD standard conditions(!) - i.e. T=60 F and p=14.696 psia
}
\examples{
 Vm_SI()
}
