% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{wd_US}
\alias{wd_US}
\title{Built-in constant: standard water density}
\format{An object of class \code{numeric} of length 1.}
\usage{
wd_US
}
\description{
Standard water density at FIELD standard conditions, 62.3664 [lb/ft3]
}
\keyword{datasets}

