% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{Psc_SI}
\alias{Psc_SI}
\title{Built-in constant: standard pressure}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
Psc_SI
}
\description{
Standard pressure is METRIC units, 1.01325 [bar]
}
\keyword{datasets}
