% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{Psc_US}
\alias{Psc_US}
\title{Standard pressure}
\format{An object of class \code{numeric} of length 1.}
\usage{
Psc_US
}
\description{
Standard pressure is FIELD units, [psia]
}
\keyword{datasets}

