% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{Tsc_SI}
\alias{Tsc_SI}
\title{Built-in constant: standard temperature}
\format{An object of class \code{numeric} of length 1.}
\usage{
Tsc_SI
}
\description{
Standard temperature is METRIC units, 293.15 [K]
}
\keyword{datasets}

