% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Units.R
\docType{data}
\name{Tsc_US}
\alias{Tsc_US}
\title{Standard temperature}
\format{An object of class \code{numeric} of length 1.}
\usage{
Tsc_US
}
\description{
Standard temperature is FIELD units, [R]
}
\keyword{datasets}

